requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
fastapi-csrf-protect==1.0.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
motor==3.7.0
packaging==24.2
passlib==1.7.4
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyJWT==2.10.1
pymongo==4.11.3
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
pytest
httpx
openai
email-validator
